:recycle: Extract calendar state
